+++ src/lib/constants.js
@@ -1,32 +1,30 @@
-
 /**
- *	constants.js: grunt-check-outdated
+ * lib/constants.js: grunt-check-outdated
  *
- *  @module grunt-check-outdated/constants
+ * @module grunt-check-outdated/constants
  *
  *//*
  *  © 2021, db-developer.
  *
- *  constants.js  is distributed  WITHOUT  ANY WARRANTY;  without  even  the
- *  implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *  Distributed  WITHOUT  ANY WARRANTY;  without  even the  implied
+ *  warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  */
 "use strict";
 
 /**
- *  Stringtable
- *  @ignore
+ *  The task name used to register the 'check_outdated' multitask.
+ *
+ *  @constant
+ *  @type {string}
+ *  @default
  */
-const _STRINGS = {
-  PROPERTY_TASKNAME_CHKOUTDATED:   "TASKNAME_NYCMOCHA",
-  PROPERTY_TASKDESC_CHKOUTDATED:   "TASKDESCRIPTION_NYCMOCHA",
-  TASKNAME_CHKOUTDATED:            "check_outdated",
-  TASKDESCRIPTION_CHKOUTDATED:     "Check package.json for outdated dependencies"
-};
+module.exports.TASKNAME = "check_outdated";
 
-// Module exports:
-Object.defineProperty( module.exports, _STRINGS.PROPERTY_TASKNAME_CHKOUTDATED, {
-  value:    _STRINGS.TASKNAME_CHKOUTDATED,
-  writable: false, enumerable: true, configurable: false });
-Object.defineProperty( module.exports, _STRINGS.PROPERTY_TASKDESC_CHKOUTDATED, {
-  value:    _STRINGS.TASKDESCRIPTION_CHKOUTDATED,
-  writable: false, enumerable: true, configurable: false });
+/**
+ *  The description used for the 'check_outdated' multitask.
+ *
+ *  @constant
+ *  @type {string}
+ *  @default
+ */
+module.exports.TASKDESCRIPTION = "Run check outdated from grunt.";