+++ src/lib/tasks/checkoutdated.js
@@ -1,44 +1,35 @@
 /**
- *	tasks/checkoutdated.js: grunt-check-outdated
+ *  tasks/checkoutdated.js: grunt-check-outdated
  *
  *  @module grunt-check-outdated/tasks/checkoutdated
  *
  *//*
- *  © 2021, slashlib.org.
+ *  © 2021, db-developer.
  *
- *  tasks/checkoutdated.js  is distributed WITHOUT ANY WARRANTY; without even the
- *  implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- *
- *//* eslint-disable-next-line */
+ *  Distributed  WITHOUT  ANY WARRANTY;  without  even the  implied
+ *  warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ */
 "use strict";
 
-/**
- *  Module initializer
- *  @ignore
- */
-const _m = {
-  const:              require( "../constants" ),
-  checkoutdatedopts:  require( "../options/checkoutdated" )
-};
+const constants = require( "../constants" );
+const options   = require( "../options/checkoutdated" );
 
 /**
  *  Stringtable initializer
  *  @ignore
  */
 function _init_STRINGS() {
-  const executecheckoutdated  = "executeChkOutdated";
-  const missingproperty       = "Missing property";
+  const execute         = "execute";
+  const missingproperty = "Missing property";
 
   return {
-    ERROR_MSG_MISSING_ARGS:         `${ executecheckoutdated }: ${ missingproperty } 'obj.args'.`,
-    ERROR_MSG_MISSING_OPTS:         `${ executecheckoutdated }: ${ missingproperty } 'obj.opts'.`,
-    ERROR_MSG_MISSING_CWD:          `${ executecheckoutdated }: ${ missingproperty } 'options.cwd'.`,
-    ERROR_MSG_MISSING_NODEEXEC:     `${ executecheckoutdated }: ${ missingproperty } 'options.node.exec'.`,
-    EXECUTECHKOUTDATED:             `${ executecheckoutdated }`,
+    ERROR_MSG_MISSING_PARAM:        `${ execute }: Missing parameter 'obj'.`,
+    ERROR_MSG_MISSING_ARGS:         `${ execute }: ${ missingproperty } 'obj.args'.`,
+    ERROR_MSG_MISSING_OPTS:         `${ execute }: ${ missingproperty } 'obj.opts'.`,
+    ERROR_MSG_MISSING_CWD:          `${ execute }: ${ missingproperty } 'options.cwd'.`,
+    ERROR_MSG_MISSING_NODEEXEC:     `${ execute }: ${ missingproperty } 'options.node.exec'.`,
     IGNORE:                         "ignore",
-    INHERIT:                        "inherit",
-    REGISTERMULTITASKCHKOUTDATED:   "registerMultiTaskChkOutdated",
-    RUNTASKCHKOUTDATED:             "runTaskChkOutdated"
+    INHERIT:                        "inherit"
   };
 }
 
@@ -49,6 +40,40 @@ function _init_STRINGS() {
 const _STRINGS = _init_STRINGS();
 
 /**
+ * Promisified adapter for `grunt.util.spawn`.
+ *
+ * Wraps the callback-based `grunt.util.spawn` API into a Promise so it can be
+ * consumed via async/await without mixing callback and business logic.
+ * The Promise resolves with the spawn result or rejects with the execution error.
+ *
+ * @param {grunt} grunt
+ *        The Grunt runtime instance providing `grunt.util.spawn`.
+ * @param {Object} config
+ *        Spawn configuration passed through to `grunt.util.spawn`.
+ * @param {string} config.cmd
+ *        Executable to run.
+ * @param {Array<string>} config.args
+ *        Command-line arguments passed to the executable.
+ * @param {Object} config.opts
+ *        Options object forwarded to the spawn call (env, cwd, stdio).
+ *
+ * @returns {Promise<Object>}
+ *          Resolves with the spawn result object provided by Grunt,
+ *          or rejects with an Error if the process execution fails.
+ */
+module.exports.spawnAsync = function spawnAsync(grunt, config) {
+  return new Promise((resolve, reject) => {
+    grunt.util.spawn(config, (error, result) => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve(result);
+      }
+    });
+  });
+}
+
+/**
  *  Return a promise for executing
  *    'node --[node opts] check-outdated --[opts]'
  *
@@ -56,57 +81,49 @@ const _STRINGS = _init_STRINGS();
  *  @param  {grunt.task}  task  the current task
  *  @param  {Object}      obj wrapper for options and arguments.
  */
-function executeChkOutdated( grunt, task, obj ) {
-  return new Promise(( resolve, reject ) => {
-    try {
-      let args = undefined;
-      if (( ! obj ) || ( ! obj.args )) {
-            throw new Error( _STRINGS.ERROR_MSG_MISSING_ARGS );
-      }
-      else args = obj.args;
+module.exports.execute = async function execute( grunt, task, obj ) {
+  if ( ! obj ) {
+       throw new Error( _STRINGS.ERROR_MSG_MISSING_PARAM );
+  }
 
-      let options = undefined;
-      if (( ! obj ) || ( ! obj.opts )) {
-            throw new Error( _STRINGS.ERROR_MSG_MISSING_OPTS );
-      }
-      else  options = obj.opts;
+  if ( ! obj.args ) {
+       throw new Error( _STRINGS.ERROR_MSG_MISSING_ARGS );
+  }
 
-      const env = process.env;
+  if ( ! obj.opts ) {
+       throw new Error( _STRINGS.ERROR_MSG_MISSING_OPTS );
+  }
 
-      let   cwd = undefined;
-      if ( ! options.cwd ) {
-           throw new Error( _STRINGS.ERROR_MSG_MISSING_CWD );
-      }
-      else cwd = options.cwd;
+  const args    = obj.args;
+  const options = obj.opts;
 
-      let   stdio = options.quiet ? _STRINGS.IGNORE : _STRINGS.INHERIT;
-      const opts  = { env, cwd, stdio };
+  if ( ! options.cwd ) {
+        throw new Error( _STRINGS.ERROR_MSG_MISSING_CWD );
+  }
 
-      let   cmd = undefined;
-      if ( ! options.node.exec ) {
-           throw new Error( _STRINGS.ERROR_MSG_MISSING_NODEEXEC );
-      }
-      else cmd = options.node.exec;
+  if ( ! options.node.exec ) {
+        throw new Error( _STRINGS.ERROR_MSG_MISSING_NODEEXEC );
+  }
 
-      const logmsg = `Will execute (stdio: '${ stdio }'): ${ cmd } ${ args.join( " " )}`;
-      /* istanbul ignore else */
-      if ( options.dryrun ) {
-           grunt.log.ok( logmsg );
-           return resolve( obj );
-      }
-      else grunt.verbose.ok( logmsg );
+  const env   = { ...process.env };
+  const cwd   = options.cwd;
+  const stdio = options.quiet ? _STRINGS.IGNORE : _STRINGS.INHERIT;
+  const opts  = { env, cwd, stdio };
+  const cmd   = options.node.exec;
 
-      /* istanbul ignore next */
-      grunt.util.spawn({ cmd, args, opts }, ( error, result ) => {
-        if ( ! error ) {
-             obj.result = result;
-             resolve( obj );
-        }
-        else reject( error );
-      });
-    }
-    catch( error ) { reject( error ); }
-  });
+  const logmsg = `Will execute (stdio: '${ stdio }'): ${ cmd } ${ args.join( " " )}`;
+  
+  /* istanbul ignore else */
+  if ( options.dryrun ) {
+        grunt.log.ok( logmsg );
+        return obj;
+  }
+  else grunt.verbose.ok( logmsg );
+
+  const result = await module.exports.spawnAsync(grunt, { cmd, args, opts });
+  obj.result = result;
+
+  return obj;
 }
 
 /**
@@ -114,38 +131,29 @@ function executeChkOutdated( grunt, task, obj ) {
  *
  *  @return {Promise} ... required by callee to terminate async call (on "then")
  */
-function runTaskChkOutdated( grunt, task ) {
-  let    promise = _m.checkoutdatedopts.toArgs( grunt, task ); // prepare args for test runs ...
-         promise = promise.then(( obj ) => { // run the tests...
-                     return executeChkOutdated( grunt, task, obj );
-                   });
-  return promise;
+module.exports.runTask = async function runTask( grunt, task ) {
+  const obj = await options.toArgs(grunt, task);
+  return module.exports.execute(grunt, task, obj);
 }
 
 /**
  *  Registers the 'check_outdated' multitask.
+ * 
+ *  Note:
+ *  - Any errors thrown in lower-level functions (e.g., `execute`, `toArgs`, `spawnAsync`)
+ *    will automatically reject the returned Promise.
+ *  - Rejected Promises are handled by logging the error and calling `done(false)` to fail
+ *    the task.* 
  *
  *  @param  {grunt} grunt
  */
-function registerMultiTaskChkOutdated( grunt ) {
-  grunt.registerMultiTask( _m.const.TASKNAME_NYCMOCHA, _m.const.TASKDESCRIPTION_NYCMOCHA,
+module.exports.registerMultiTask = function registerMultiTask( grunt ) {
+  grunt.registerMultiTask( constants.TASKNAME, constants.TASKDESCRIPTION,
     /* istanbul ignore next */ function () {
       const task = this;
       const done = task.async();
-      runTaskChkOutdated( grunt, task ).then((       ) => { done(); },
-                                             ( error ) => { grunt.log.error( error ); done( false ); });
+      module.exports.runTask( grunt, task )
+            .then((       ) => { done(); },
+                  ( error ) => { grunt.log.error( error ); done( false ); });
   });
 }
-
-/* eslint-disable */
-// Module exports:
-Object.defineProperty( module.exports, _STRINGS.EXECUTECHKOUTDATED,  {
-       value:    executeChkOutdated,
-       writable: false, enumerable: true, configurable: false });
-Object.defineProperty( module.exports, _STRINGS.REGISTERMULTITASKCHKOUTDATED, {
-       value:    registerMultiTaskChkOutdated,
-       writable: false, enumerable: true, configurable: false });
-Object.defineProperty( module.exports, _STRINGS.RUNTASKCHKOUTDATED,  {
-       value:    runTaskChkOutdated,
-       writable: false, enumerable: true, configurable: false });
-/* eslint-enable */