+++ src/lib/tasks/index.js
@@ -1,47 +1,26 @@
 /**
- *	index.js: grunt-check-outdated/tasks
+ *	tasks/index.js: grunt-check-outdated/tasks
  *
  *  @module grunt-check-outdated/tasks
  *
  *//*
- *  © 2021, slashlib.org.
+ *  © 2021, db-developer.
  *
- *  index.js  is distributed WITHOUT ANY WARRANTY; without even the implied
- *  warranty  of  MERCHANTABILITY  or  FITNESS  FOR  A PARTICULAR  PURPOSE.
- *
+ *  Distributed  WITHOUT  ANY WARRANTY;  without  even the  implied
+ *  warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  */
 "use strict";
 
-/**
- *  Module initializer
- *  @ignore
- */
-const _m = {
-  checkoutdated:     require( "./checkoutdated" )
-};
+const checkoutdated = require( "./checkoutdated" );
 
 /**
- *  Stringtable
- *  @ignore
- */
-const _STRINGS = {
-  REGISTERMULTITASKCHKOUTDATED:  "registerMultiTaskChkOutdated",
-  RUNTASKCHKOUTDATED:            "runTaskChkOutdated"
-};
-
-// Module exports:
-/**
- *  Register a multitask for check_outdated.
+ *  Registers the `check_outdated` multitask with Grunt.
  *
- *  @see    Function [registerMultiTaskChkOutdated]{@link chkoutdated.md#.registerMultiTaskChkOutdated}
- *          published by module chkoutdated for a detailed function description.
+ *  This function is the main entry point to integrate `grunt-call-rollup`
+ *  into a Gruntfile. Internally, it forwards to
+ *  [chkoutdated.registerMultiTask]{@link ./chkoutdated.md#.registerMultiTask}.
  *
- *  @function module:grunt-check-outdated/tasks.registerMultiTaskChkOutdated
- *  @param  {grunt} grunt
+ *  @function module:grunt-check-outdated/tasks.registerMultiTask
+ *  @param  {grunt} grunt - The Grunt runtime instance.
  */
-Object.defineProperty( module.exports, _STRINGS.REGISTERMULTITASKCHKOUTDATED, {
-  value:    _m.checkoutdated.registerMultiTaskChkOutdated,
-  writable: false, enumerable: true, configurable: false });
-Object.defineProperty( module.exports, _STRINGS.RUNTASKCHKOUTDATED,  {
-  value:    _m.checkoutdated.runTaskChkOutdated,
-  writable: false, enumerable: true, configurable: false });
+module.exports.registerMultiTask = checkoutdated.registerMultiTask